Change non-human being part of max text
The text that said "non-human being" is correct for max himself, but we
use pictures of the hacker, who is human....or is he? So the text is
changed to wave catcher. This has two meanings: first he says "catch the
wave" in the signal highjacking, secondly he did highjack the signal by
sending higher frequency waves to the viewers than the tv signal.